client: remote players froze because render time was held in a float
Two players standing next to each other could see one another's bodies at the
point they spawned and no movement afterwards, while thirty packets a second
arrived, the clocks agreed to within twenty milliseconds and the appearance
applied correctly.

Everything looked healthy because everything WAS healthy except one conversion.

InterpolateEntity began:

    const float tick = NetworkWorldSystem::GetTick() - aSimulationDelay;

GetTick() returns milliseconds since the epoch as a uint64 - about 1.787e12. A
32-bit float carries 24 bits of mantissa, so at that magnitude the gap between
representable values is 131072, or roughly 131 seconds. Converting the tick to
float discarded everything below that, three times over:

1. Subtracting the 100ms simulation delay produced a bit-identical result, so
   the interpolation buffer effectively had no delay.
2. In the drain loop a sample 30ms old and current render time rounded to the
   same float, so `sample.Tick <= tick` held for the entire buffer and it
   emptied every frame.
3. With the buffer always empty the extrapolation branch computed
   `ahead = tick - first.Tick` as exactly 0.0, hit its own `ahead <= 0` guard,
   and returned without applying anything - every frame, permanently.

So the puppet stayed wherever it was last placed, which is its spawn point.

Verified against real ticks taken from a session log: float(1787286555346) and
float(1787286555346 - 100) are the same number, and the drain comparison, the
extrapolation delta and the interpolation ratio all collapse to zero.

The fix is the same everywhere: absolute ticks stay in int64, and only
DIFFERENCES - tens of milliseconds - are allowed to become float. That covers
render time, the drain comparison, the extrapolation window, the segment ratio
and LastRenderTick, which was a float holding an absolute tick for the same
reason.

Two other freeze paths appear in the same logs and are NOT addressed here: a
spawn race where movement arrives before the puppet is registered under that id
(self-heals in about 25ms), and the entity-stub gate. Both are intermittent and
both log. This one fired silently for every remote player, always.

Compiles. NOT yet verified in a live two-player session - that is the next step.

+++ code/client/App/Components/InterpolationComponent.h
@@ -32,5 +32,11 @@ struct InterpolationComponent
 
     // Render time as of the last frame. The vehicle path needs a frame delta, which used
     // to fall out of PreviousFrame back when that was rewritten every frame.
-    float LastRenderTick{0.f};
+    //
+    // int64, not float. Ticks are milliseconds since the epoch - around 1.787e12 - and a
+    // 32-bit float has 24 bits of mantissa, so consecutive representable values that far
+    // out are 131072 apart. Every tick within the same two-minute window collapsed onto
+    // one value, which made every comparison against render time meaningless. See the
+    // note in InterpolateEntity.
+    int64_t LastRenderTick{0};
 };
